create extra_field_repository
